+++ package.json
@@ -31,8 +31,7 @@
     "doctor": "node ./packages/doctor",
     "e2e-test": "lerna exec \"gulp e2e-test\"",
     "generate-docs": "lerna run --scope=appium generate-docs",
-    "pregenerate-schema-declarations": "lerna exec --scope=appium \"gulp transpile\"",
-    "generate-schema-declarations": "node ./scripts/generate-schema-declarations.js",
+    "generate-schema-declarations": "node ./scripts/generate-schema-declarations.mjs",
     "postinstall": "lerna bootstrap && lerna exec --parallel \"gulp prepublish\"",
     "install-fake-driver": "lerna run --scope=appium install-fake-driver",
     "lint": "lerna exec --parallel \"gulp lint\" && lerna run --scope=@appium/eslint-config-appium lint",
